{"schema":"decodex/commit/1","summary":"Split search ranking scope boost helpers after strict validation.","authority":"manual"}
