+++ packages/elf-service/src/search/ranking/text.rs
@@ -1,10 +1,11 @@
+mod scope;
+
 use std::collections::{HashMap, HashSet};
 
 use time::OffsetDateTime;
 
 use crate::search::DeterministicRankingTerms;
 use elf_config::{Config, Context};
-use elf_domain::english_gate;
 
 pub fn build_dense_embedding_input(
 	query: &str,
@@ -24,75 +25,7 @@ pub fn build_scope_context_boost_by_scope<'a>(
 	tokens: &[String],
 	context: Option<&'a Context>,
 ) -> HashMap<&'a str, f32> {
-	let Some(context) = context else { return HashMap::new() };
-	let Some(weight) = context.scope_boost_weight else { return HashMap::new() };
-
-	if weight <= 0.0 || tokens.is_empty() {
-		return HashMap::new();
-	}
-
-	let Some(descriptions) = context.scope_descriptions.as_ref() else { return HashMap::new() };
-	let mut out = HashMap::new();
-
-	for (scope, description) in descriptions {
-		let boost = scope_description_boost(tokens, description, weight);
-
-		if boost > 0.0 {
-			out.insert(scope.as_str(), boost);
-		}
-	}
-
-	out
-}
-
-pub fn scope_description_boost(tokens: &[String], description: &str, weight: f32) -> f32 {
-	if weight <= 0.0 || tokens.is_empty() {
-		return 0.0;
-	}
-
-	let trimmed = description.trim();
-
-	if trimmed.is_empty() || !english_gate::is_english_natural_language(trimmed) {
-		return 0.0;
-	}
-
-	let mut normalized = String::with_capacity(trimmed.len());
-
-	for ch in trimmed.chars() {
-		if ch.is_ascii_alphanumeric() {
-			normalized.push(ch.to_ascii_lowercase());
-		} else {
-			normalized.push(' ');
-		}
-	}
-
-	let mut description_tokens = HashSet::new();
-
-	for token in normalized.split_whitespace() {
-		if token.len() < 2 {
-			continue;
-		}
-
-		description_tokens.insert(token);
-	}
-
-	if description_tokens.is_empty() {
-		return 0.0;
-	}
-
-	let mut matched = 0_usize;
-
-	for token in tokens {
-		if description_tokens.contains(token.as_str()) {
-			matched += 1;
-		}
-	}
-
-	if matched == 0 {
-		return 0.0;
-	}
-
-	weight * (matched as f32 / tokens.len() as f32)
+	scope::build_scope_context_boost_by_scope(tokens, context)
 }
 
 pub fn tokenize_query(query: &str, max_terms: usize) -> Vec<String> {
